+++ src/renderer/svg-components/index.ts
@@ -4,3 +4,4 @@ export * from './select-arrow-right.component';
 export * from './select-arrow-down.component';
 export * from './refresh-images.component';
 export * from './image-alert.component';
+export * from './add-images.component';